flask-app/requirements.txt
Flask==2.3.2
pymongo==4.6.0
gridfs==1.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
langchain-openai==0.0.2 
reportlab==3.6.12
